requirements.txt
google-genai
ollama
requests
beautifulsoup4
loguru
Pillow
lxml
